HDDS-9456. Non-blocking cache for Container State Machine

+++ hadoop-hdds/common/src/main/java/org/apache/hadoop/hdds/utils/ResourceCache.java
@@ -32,7 +32,7 @@ public class ResourceCache<K, V> implements Cache<K, V> {
   private final com.google.common.cache.Cache<K, V> cache;
 
   public ResourceCache(
-      Weigher weigher, int limits,
+      Weigher weigher, long limits,
       RemovalListener<K, V> listener) {
     Objects.requireNonNull(weigher);
     if (listener == null) {
